+++ src/builder/arg.rs
@@ -14,6 +14,7 @@ use std::{env, ffi::OsString};
 // Internal
 use super::{ArgFlags, ArgSettings};
 use crate::builder::ArgPredicate;
+use crate::builder::ValuesRange;
 use crate::util::{Id, Key};
 use crate::ArgAction;
 use crate::PossibleValue;
@@ -72,9 +73,7 @@ pub struct Arg<'help> {
     pub(crate) short_aliases: Vec<(char, bool)>, // (name, visible)
     pub(crate) disp_ord: Option<usize>,
     pub(crate) val_names: Vec<&'help str>,
-    pub(crate) num_vals: Option<usize>,
-    pub(crate) max_vals: Option<usize>,
-    pub(crate) min_vals: Option<usize>,
+    pub(crate) num_vals: Option<ValuesRange>,
     pub(crate) val_delim: Option<char>,
     pub(crate) default_vals: Vec<&'help OsStr>,
     pub(crate) default_vals_ifs: Vec<(Id, ArgPredicate<'help>, Option<&'help OsStr>)>,
@@ -1066,9 +1065,9 @@ impl<'help> Arg<'help> {
     ///
     /// [`subcommands`]: crate::Command::subcommand()
     /// [`Arg::number_of_values(1)`]: Arg::number_of_values()
-    /// [maximum number of values]: Arg::max_values()
+    /// [maximum number of values]: Arg::number_of_values()
     /// [specific number of values]: Arg::number_of_values()
-    /// [maximum]: Arg::max_values()
+    /// [maximum]: Arg::number_of_values()
     /// [specific]: Arg::number_of_values()
     #[inline]
     #[must_use]
@@ -1080,166 +1079,160 @@ impl<'help> Arg<'help> {
         }
     }
 
-    /// The number of values allowed for this argument.
+    /// Specifies the number of values allowed per occurrence of this argument
     ///
-    /// For example, if you had a
-    /// `-f <file>` argument where you wanted exactly 3 'files' you would set
-    /// `.number_of_values(3)`, and this argument wouldn't be satisfied unless the user provided
-    /// 3 and only 3 values.
+    /// For example, if you had a `-f <file>` argument where you wanted exactly 3 'files' you would
+    /// set `.number_of_values(3)`, and this argument wouldn't be satisfied unless the user
+    /// provided 3 and only 3 values.
     ///
-    /// **NOTE:** implicitly sets [`Arg::action(ArgAction::Set)`] and [`Arg::multiple_values(true)`].
+    /// **NOTE:** Users may specify values for arguments in any of the following methods
+    ///
+    /// - Using a space such as `-o value` or `--option value`
+    /// - Using an equals and no space such as `-o=value` or `--option=value`
+    /// - Use a short and no space such as `-ovalue`
+    ///
+    /// **WARNING:**
+    ///
+    /// Setting a variable number of values (e.g. `1..=10`) for an argument without
+    /// other details can be dangerous in some circumstances. Because multiple values are
+    /// allowed, `--option val1 val2 val3` is perfectly valid. Be careful when designing a CLI
+    /// where **positional arguments** or **subcommands** are *also* expected as `clap` will continue
+    /// parsing *values* until one of the following happens:
+    ///
+    /// - It reaches the maximum number of values
+    /// - It reaches a specific number of values
+    /// - It finds another flag or option (i.e. something that starts with a `-`)
+    /// - It reaches the [`Arg::value_terminator`] if set
+    ///
+    /// Alternatively,
+    /// - Require a delimiter between values with [Arg::require_value_delimiter]
+    /// - Require a flag occurrence per value with [`ArgAction::Append`]
+    /// - Require positional arguments to appear after `--` with [`Arg::last`]
     ///
     /// # Examples
     ///
+    /// Option:
     /// ```rust
     /// # use clap::{Command, Arg};
-    /// Arg::new("file")
-    ///     .short('f')
-    ///     .number_of_values(3);
+    /// let m = Command::new("prog")
+    ///     .arg(Arg::new("mode")
+    ///         .long("mode")
+    ///         .number_of_values(1))
+    ///     .get_matches_from(vec![
+    ///         "prog", "--mode", "fast"
+    ///     ]);
+    ///
+    /// assert_eq!(m.get_one::<String>("mode").unwrap(), "fast");
     /// ```
     ///
-    /// Not supplying the correct number of values is an error
-    ///
+    /// Flag/option hybrid (see also [default_missing_value][Arg::default_missing_value])
     /// ```rust
     /// # use clap::{Command, Arg, ErrorKind, ArgAction};
-    /// let res = Command::new("prog")
+    /// let cmd = Command::new("prog")
+    ///     .arg(Arg::new("mode")
+    ///         .long("mode")
+    ///         .default_missing_value("slow")
+    ///         .default_value("plaid")
+    ///         .number_of_values(0..=1));
+    ///
+    /// let m = cmd.clone()
+    ///     .get_matches_from(vec![
+    ///         "prog", "--mode", "fast"
+    ///     ]);
+    /// assert_eq!(m.get_one::<String>("mode").unwrap(), "fast");
+    ///
+    /// let m = cmd.clone()
+    ///     .get_matches_from(vec![
+    ///         "prog", "--mode",
+    ///     ]);
+    /// assert_eq!(m.get_one::<String>("mode").unwrap(), "slow");
+    ///
+    /// let m = cmd.clone()
+    ///     .get_matches_from(vec![
+    ///         "prog",
+    ///     ]);
+    /// assert_eq!(m.get_one::<String>("mode").unwrap(), "plaid");
+    /// ```
+    ///
+    /// Tuples
+    /// ```rust
+    /// # use clap::{Command, Arg, ErrorKind, ArgAction};
+    /// let cmd = Command::new("prog")
     ///     .arg(Arg::new("file")
     ///         .action(ArgAction::Set)
     ///         .number_of_values(2)
-    ///         .short('F'))
+    ///         .short('F'));
+    ///
+    /// let m = cmd.clone()
+    ///     .get_matches_from(vec![
+    ///         "prog", "-F", "in-file", "out-file"
+    ///     ]);
+    /// assert_eq!(
+    ///     m.get_many::<String>("file").unwrap_or_default().map(|v| v.as_str()).collect::<Vec<_>>(),
+    ///     vec!["in-file", "out-file"]
+    /// );
+    ///
+    /// let res = cmd.clone()
     ///     .try_get_matches_from(vec![
     ///         "prog", "-F", "file1"
     ///     ]);
-    ///
-    /// assert!(res.is_err());
     /// assert_eq!(res.unwrap_err().kind(), ErrorKind::WrongNumberOfValues);
     /// ```
-    #[inline]
-    #[must_use]
-    pub fn number_of_values(mut self, qty: usize) -> Self {
-        self.num_vals = Some(qty);
-        self.takes_value(qty != 0).multiple_values(1 < qty)
-    }
-
-    /// The *maximum* number of values are for this argument.
     ///
-    /// For example, if you had a
-    /// `-f <file>` argument where you wanted up to 3 'files' you would set `.max_values(3)`, and
-    /// this argument would be satisfied if the user provided, 1, 2, or 3 values.
-    ///
-    /// # Examples
-    ///
-    /// ```rust
-    /// # use clap::{Command, Arg};
-    /// Arg::new("file")
-    ///     .short('f')
-    ///     .max_values(3);
-    /// ```
-    ///
-    /// Supplying less than the maximum number of values is allowed
-    ///
+    /// A common mistake is to define an option which allows multiple values and a positional
+    /// argument.
     /// ```rust
     /// # use clap::{Command, Arg, ArgAction};
-    /// let res = Command::new("prog")
+    /// let cmd = Command::new("prog")
     ///     .arg(Arg::new("file")
     ///         .action(ArgAction::Set)
-    ///         .max_values(3)
+    ///         .number_of_values(0..)
     ///         .short('F'))
-    ///     .try_get_matches_from(vec![
-    ///         "prog", "-F", "file1", "file2"
-    ///     ]);
+    ///     .arg(Arg::new("word"));
     ///
-    /// assert!(res.is_ok());
-    /// let m = res.unwrap();
+    /// let m = cmd.clone().get_matches_from(vec![
+    ///     "prog", "-F", "file1", "file2", "file3", "word"
+    /// ]);
     /// let files: Vec<_> = m.get_many::<String>("file").unwrap().collect();
-    /// assert_eq!(files, ["file1", "file2"]);
-    /// ```
+    /// assert_eq!(files, ["file1", "file2", "file3", "word"]); // wait...what?!
+    /// assert!(!m.contains_id("word")); // but we clearly used word!
     ///
-    /// Supplying more than the maximum number of values is an error
-    ///
-    /// ```rust
-    /// # use clap::{Command, Arg, ErrorKind, ArgAction};
-    /// let res = Command::new("prog")
-    ///     .arg(Arg::new("file")
-    ///         .action(ArgAction::Set)
-    ///         .max_values(2)
-    ///         .short('F'))
-    ///     .try_get_matches_from(vec![
-    ///         "prog", "-F", "file1", "file2", "file3"
-    ///     ]);
-    ///
-    /// assert!(res.is_err());
-    /// assert_eq!(res.unwrap_err().kind(), ErrorKind::UnknownArgument);
-    /// ```
-    #[inline]
-    #[must_use]
-    pub fn max_values(mut self, qty: usize) -> Self {
-        self.max_vals = Some(qty);
-        self.takes_value(true).multiple_values(true)
-    }
-
-    /// The *minimum* number of values for this argument.
-    ///
-    /// For example, if you had a
-    /// `-f <file>` argument where you wanted at least 2 'files' you would set
-    /// `.min_values(2)`, and this argument would be satisfied if the user provided, 2 or more
-    /// values.
-    ///
-    /// **NOTE:** Passing a non-zero value is not the same as specifying [`Arg::required(true)`].
-    /// This is due to min and max validation only being performed for present arguments,
-    /// marking them as required will thus perform validation and a min value of 1
-    /// is unnecessary, ignored if not required.
-    ///
-    /// # Examples
-    ///
-    /// ```rust
-    /// # use clap::{Command, Arg};
-    /// Arg::new("file")
-    ///     .short('f')
-    ///     .min_values(3);
-    /// ```
-    ///
-    /// Supplying more than the minimum number of values is allowed
-    ///
-    /// ```rust
-    /// # use clap::{Command, Arg, ArgAction};
-    /// let res = Command::new("prog")
-    ///     .arg(Arg::new("file")
-    ///         .action(ArgAction::Set)
-    ///         .min_values(2)
-    ///         .short('F'))
-    ///     .try_get_matches_from(vec![
-    ///         "prog", "-F", "file1", "file2", "file3"
-    ///     ]);
-    ///
-    /// assert!(res.is_ok());
-    /// let m = res.unwrap();
+    /// // but this works
+    /// let m = cmd.clone().get_matches_from(vec![
+    ///     "prog", "word", "-F", "file1", "file2", "file3",
+    /// ]);
     /// let files: Vec<_> = m.get_many::<String>("file").unwrap().collect();
     /// assert_eq!(files, ["file1", "file2", "file3"]);
+    /// assert_eq!(m.get_one::<String>("word").unwrap(), "word");
     /// ```
+    /// The problem is `clap` doesn't know when to stop parsing values for "file".
     ///
-    /// Supplying less than the minimum number of values is an error
-    ///
+    /// A solution for the example above is to limit how many values with a maximum, or specific
+    /// number, or to say [`ArgAction::Append`] is ok, but multiple values are not.
     /// ```rust
-    /// # use clap::{Command, Arg, ErrorKind, ArgAction};
-    /// let res = Command::new("prog")
+    /// # use clap::{Command, Arg, ArgAction};
+    /// let m = Command::new("prog")
     ///     .arg(Arg::new("file")
-    ///         .action(ArgAction::Set)
-    ///         .min_values(2)
+    ///         .action(ArgAction::Append)
     ///         .short('F'))
-    ///     .try_get_matches_from(vec![
-    ///         "prog", "-F", "file1"
+    ///     .arg(Arg::new("word"))
+    ///     .get_matches_from(vec![
+    ///         "prog", "-F", "file1", "-F", "file2", "-F", "file3", "word"
     ///     ]);
     ///
-    /// assert!(res.is_err());
-    /// assert_eq!(res.unwrap_err().kind(), ErrorKind::TooFewValues);
+    /// let files: Vec<_> = m.get_many::<String>("file").unwrap().collect();
+    /// assert_eq!(files, ["file1", "file2", "file3"]);
+    /// assert_eq!(m.get_one::<String>("word").unwrap(), "word");
     /// ```
-    /// [`Arg::required(true)`]: Arg::required()
+    /// [`Arg::value_delimiter(char)`]: Arg::value_delimiter()
+    /// [multiple values]: Arg::multiple_values
     #[inline]
     #[must_use]
-    pub fn min_values(mut self, qty: usize) -> Self {
-        self.min_vals = Some(qty);
-        self.takes_value(true).multiple_values(true)
+    pub fn number_of_values(mut self, qty: impl Into<ValuesRange>) -> Self {
+        let qty = qty.into();
+        self.num_vals = Some(qty);
+        self.takes_value(qty.takes_values())
+            .multiple_values(qty.is_multiple())
     }
 
     /// Placeholder for the argument's value in the help message / usage.
@@ -1740,8 +1733,7 @@ impl<'help> Arg<'help> {
     /// By default when
     /// one sets [`multiple_values(true)`] on an argument, clap will continue parsing values for that
     /// argument until it reaches another valid argument, or one of the other more specific settings
-    /// for multiple values is used (such as [`min_values`], [`max_values`] or
-    /// [`number_of_values`]).
+    /// for multiple values is used (such as [`number_of_values`]).
     ///
     /// **NOTE:** This setting only applies to [options] and [positional arguments]
     ///
@@ -1781,9 +1773,7 @@ impl<'help> Arg<'help> {
     /// [options]: Arg::takes_value()
     /// [positional arguments]: Arg::index()
     /// [`multiple_values(true)`]: Arg::multiple_values()
-    /// [`min_values`]: Arg::min_values()
     /// [`number_of_values`]: Arg::number_of_values()
-    /// [`max_values`]: Arg::max_values()
     #[inline]
     #[must_use]
     pub fn value_terminator(mut self, term: &'help str) -> Self {
@@ -1931,9 +1921,10 @@ impl<'help> Arg<'help> {
     /// argument is a common example. By, supplying an default, such as `default_missing_value("always")`,
     /// the user can quickly just add `--color` to the command line to produce the desired color output.
     ///
-    /// **NOTE:** using this configuration option requires the use of the `.min_values(0)` and the
-    /// `.require_equals(true)` configuration option. These are required in order to unambiguously
-    /// determine what, if any, value was supplied for the argument.
+    /// **NOTE:** using this configuration option requires the use of the
+    /// [`.number_of_values(0..N)`][Arg::number_of_values] and the
+    /// [`.require_equals(true)`][Arg::require_equals] configuration option. These are required in
+    /// order to unambiguously determine what, if any, value was supplied for the argument.
     ///
     /// # Examples
     ///
@@ -1946,7 +1937,7 @@ impl<'help> Arg<'help> {
     ///             .value_name("WHEN")
     ///             .value_parser(["always", "auto", "never"])
     ///             .default_value("auto")
-    ///             .min_values(0)
+    ///             .number_of_values(0..=1)
     ///             .require_equals(true)
     ///             .default_missing_value("always")
     ///             .help("Specify WHEN to colorize output.")
@@ -1983,7 +1974,7 @@ impl<'help> Arg<'help> {
     ///         .arg(Arg::new("create").long("create")
     ///             .value_name("BOOL")
     ///             .value_parser(value_parser!(bool))
-    ///             .min_values(0)
+    ///             .number_of_values(0..=1)
     ///             .require_equals(true)
     ///             .default_missing_value("true")
     ///         )
@@ -4062,13 +4053,13 @@ impl<'help> Arg<'help> {
 
     /// Get the number of values for this argument.
     #[inline]
-    pub fn get_num_vals(&self) -> Option<usize> {
+    pub fn get_num_vals(&self) -> Option<ValuesRange> {
         self.num_vals
     }
 
     #[inline]
     pub(crate) fn get_min_vals(&self) -> Option<usize> {
-        self.min_vals
+        self.get_num_vals().map(|r| r.min_values())
     }
 
     /// Get the delimiter between multiple values
@@ -4330,13 +4321,9 @@ impl<'help> Arg<'help> {
         }
 
         let val_names_len = self.val_names.len();
-
         if val_names_len > 1 {
             self.settings.set(ArgSettings::MultipleValues);
-
-            if self.num_vals.is_none() {
-                self.num_vals = Some(val_names_len);
-            }
+            self.num_vals.get_or_insert(val_names_len.into());
         }
     }
 
@@ -4497,8 +4484,6 @@ impl<'help> fmt::Debug for Arg<'help> {
             .field("disp_ord", &self.disp_ord)
             .field("val_names", &self.val_names)
             .field("num_vals", &self.num_vals)
-            .field("max_vals", &self.max_vals)
-            .field("min_vals", &self.min_vals)
             .field("val_delim", &self.val_delim)
             .field("default_vals", &self.default_vals)
             .field("default_vals_ifs", &self.default_vals_ifs)
@@ -4552,10 +4537,17 @@ pub(crate) fn render_arg_val(arg: &Arg) -> String {
 
     let mut extra_values = false;
     debug_assert!(arg.is_takes_value_set());
-    let num_vals = arg.num_vals.unwrap_or(1);
+    let num_vals = arg.num_vals.unwrap_or_else(|| {
+        if arg.is_multiple_values_set() {
+            (1..).into()
+        } else {
+            1.into()
+        }
+    });
     if val_names.len() == 1 {
         let arg_name = format!("<{}>", val_names[0]);
-        for n in 1..=num_vals {
+        let min = num_vals.min_values().max(1);
+        for n in 1..=min {
             if n != 1 {
                 rendered.push_str(delim);
             }
@@ -4571,8 +4563,9 @@ pub(crate) fn render_arg_val(arg: &Arg) -> String {
             }
             rendered.push_str(&arg_name);
         }
-        extra_values |= val_names.len() < num_vals;
+        extra_values |= val_names.len() < num_vals.max_values();
     }
+    extra_values |= !num_vals.is_fixed();
     if arg.is_positional() {
         if matches!(*arg.get_action(), ArgAction::Append) {
             extra_values = true;
@@ -4678,7 +4671,7 @@ mod test {
         let mut o = Arg::new("opt")
             .long("option")
             .action(ArgAction::Set)
-            .min_values(0);
+            .number_of_values(0..);
         o._build();
 
         assert_eq!(o.to_string(), "--option [<opt>...]");
@@ -4689,7 +4682,7 @@ mod test {
         let mut o = Arg::new("opt")
             .long("option")
             .action(ArgAction::Set)
-            .min_values(1);
+            .number_of_values(1..);
         o._build();
 
         assert_eq!(o.to_string(), "--option <opt>...");
@@ -4700,7 +4693,7 @@ mod test {
         let mut o = Arg::new("opt")
             .short('o')
             .action(ArgAction::Set)
-            .min_values(0)
+            .number_of_values(0..)
             .value_names(&["file"]);
         o._build();
 
@@ -4712,7 +4705,7 @@ mod test {
         let mut o = Arg::new("opt")
             .short('o')
             .action(ArgAction::Set)
-            .min_values(1)
+            .number_of_values(1..)
             .value_names(&["file"]);
         o._build();
 
@@ -4800,7 +4793,7 @@ mod test {
 
     #[test]
     fn positional_display_zero_or_more_values() {
-        let mut p = Arg::new("pos").index(1).min_values(0);
+        let mut p = Arg::new("pos").index(1).number_of_values(0..);
         p._build();
 
         assert_eq!(p.to_string(), "[<pos>...]");
@@ -4808,7 +4801,7 @@ mod test {
 
     #[test]
     fn positional_display_one_or_more_values() {
-        let mut p = Arg::new("pos").index(1).min_values(1);
+        let mut p = Arg::new("pos").index(1).number_of_values(1..);
         p._build();
 
         assert_eq!(p.to_string(), "<pos>...");